requirements.txt
# Core dependencies
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-amazon==8.13.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==1.4.50

# Data processing
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1

# Data generation
Faker==20.1.0

# AWS/S3/MinIO
boto3==1.29.7
s3fs==2023.12.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Data quality
great-expectations==0.18.5

# Visualization (optional)
matplotlib==3.8.2
seaborn==0.13.0

# Testing
pytest==7.4.3
